library_project/requirements.txt
Django==5.0.3
graphene-django==3.2.0
django-graphql-jwt==0.4.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-filter==24.1
gunicorn==21.2.0
